+++ DSCResources/MSFT_xCluster/MSFT_xCluster.psm1
@@ -10,6 +10,18 @@ $script:localizedData = Get-LocalizedData -ResourceName 'MSFT_xCluster'
     .PARAMETER Name
         Name of the failover cluster.
 
+    .PARAMETER StaticIPAddress
+        Static IP Address of the failover cluster.
+
+    .PARAMETER IgnoreNetwork
+        One or more networks to ignore when creating the cluster. Only networks
+        using Static IP can be ignored, networks that are assigned an IP address
+        through DHCP cannot be ignored, and are added for cluster communication.
+        To remove networks assigned an IP address through DHCP use the resource
+        xClusterNetwork to change the role of the network.
+        This parameter is only used during the creation of the cluster and is
+        not monitored after.
+
     .PARAMETER DomainAdministratorCredential
         Credential used to create the failover cluster in Active Directory.
 #>
@@ -21,6 +33,14 @@ function Get-TargetResource
         [Parameter(Mandatory = $true)]
         [System.String]
         $Name,
+
+        [Parameter()]
+        [System.String]
+        $StaticIPAddress,
+
+        [Parameter()]
+        [System.String[]]
+        $IgnoreNetwork,
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
@@ -63,6 +83,7 @@ function Get-TargetResource
     @{
         Name                          = $Name
         StaticIPAddress               = $address.Value
+        IgnoreNetwork                 = $IgnoreNetwork
         DomainAdministratorCredential = $DomainAdministratorCredential
     }
 }
@@ -76,6 +97,15 @@ function Get-TargetResource
 
     .PARAMETER StaticIPAddress
         Static IP Address of the failover cluster.
+
+    .PARAMETER IgnoreNetwork
+        One or more networks to ignore when creating the cluster. Only networks
+        using Static IP can be ignored, networks that are assigned an IP address
+        through DHCP cannot be ignored, and are added for cluster communication.
+        To remove networks assigned an IP address through DHCP use the resource
+        xClusterNetwork to change the role of the network.
+        This parameter is only used during the creation of the cluster and is
+        not monitored after.
 
     .PARAMETER DomainAdministratorCredential
         Credential used to create the failover cluster in Active Directory.
@@ -100,6 +130,10 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $StaticIPAddress,
+
+        [Parameter()]
+        [System.String[]]
+        $IgnoreNetwork,
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
@@ -139,16 +173,31 @@ function Set-TargetResource
         {
             Write-Verbose -Message ($script:localizedData.ClusterAbsent -f $Name)
 
-            if ($StaticIPAddress)
-            {
-                New-Cluster -Name $Name -Node $env:COMPUTERNAME -StaticAddress $StaticIPAddress -NoStorage -Force -ErrorAction Stop
-            }
-            else
-            {
-                New-Cluster -Name $Name -Node $env:COMPUTERNAME -NoStorage -Force -ErrorAction Stop
+            $newClusterParameters = @{
+              Name          = $Name
+              Node          = $env:COMPUTERNAME
+              NoStorage     = $true
+              Force         = $true
+              ErrorAction   = 'Stop'
             }
 
-            if ( -not (Get-Cluster))
+            if ($StaticIPAddress)
+            {
+                $newClusterParameters += @{
+                    StaticAddress = $StaticIPAddress
+                  }
+            }
+
+            if ($PSBoundParameters.ContainsKey('IgnoreNetwork'))
+            {
+                $newClusterParameters += @{
+                    IgnoreNetwork = $IgnoreNetwork
+                }
+            }
+
+            New-Cluster @newClusterParameters
+
+            if (-not (Get-Cluster))
             {
                 $errorMessage = $script:localizedData.FailedCreatingCluster
                 New-InvalidOperationException -Message $errorMessage
@@ -204,6 +253,18 @@ function Set-TargetResource
         Static IP Address of the failover cluster.
         Not used in Test-TargetResource.
 
+    .PARAMETER IgnoreNetwork
+        One or more networks to ignore when creating the cluster. Only networks
+        using Static IP can be ignored, networks that are assigned an IP address
+        through DHCP cannot be ignored, and are added for cluster communication.
+        To remove networks assigned an IP address through DHCP use the resource
+        xClusterNetwork to change the role of the network.
+        This parameter is only used during the creation of the cluster and is
+        not monitored after.
+
+        Not used in Test-TargetResource. Currently it is unknown how to determine
+        which networks are ignored, to test so they ar ein desired state.
+
     .PARAMETER DomainAdministratorCredential
         Credential used to create the failover cluster in Active Directory.
 
@@ -231,6 +292,10 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $StaticIPAddress,
+
+        [Parameter()]
+        [System.String[]]
+        $IgnoreNetwork,
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]